requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy==1.0.1
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
openai==0.27.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
protobuf==3.20.3
python-dotenv==0.21.1
requests==2.28.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
